requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
python-dotenv
fuzzywuzzy
python-Levenshtein
requests
python-multipart
bcrypt
jinja2